+++ src/app/services/cmd/system/exit.js
