+++ src/app/services/cmd/system/help.js
